+++ lua/mini/completion.lua
@@ -252,8 +252,8 @@ MiniCompletion.config = {
   -- - `height` and `width` are maximum dimensions.
   -- - `border` defines border (as in `nvim_open_win()`).
   window = {
-    info = { height = 25, width = 80, border = 'none' },
-    signature = { height = 25, width = 80, border = 'none' },
+    info = { height = 25, width = 80, border = 'single' },
+    signature = { height = 25, width = 80, border = 'single' },
   },
 
   -- Way of how module does LSP completion