+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Truncate title/footer from left if it is too wide.
     - Set default title if window is allowed to have border.
     - BREAKING: Use single space padding for default title/footer.
+    - BREAKING: Use 'single' as default window border.
 
 ## mini.ai